set chrome driver class as global utils
