+++ scrapers/news_scraper/utils/chrome_driver.py
@@ -1,0 +1,1 @@
+../../chrome_driver.py
